+++ client/src/pages/RushGame/index.tsx
@@ -1,7 +1,6 @@
 import { useEffect } from "react";
 import { motion } from "framer-motion";
 import { useCookies } from "react-cookie";
-import { RushAPI } from "@/apis/rushAPI.ts";
 import CTAButton from "@/components/CTAButton";
 import { CARD_PHASE } from "@/constants/Rush/rushCard.ts";
 import { ASCEND, SCROLL_MOTION } from "@/constants/animation.ts";
@@ -11,12 +10,10 @@ import Countdown from "@/features/RushGame/RushGameSections/Countdown.tsx";
 import FinalResult from "@/features/RushGame/RushGameSections/FinalResult.tsx";
 import SelectedCard from "@/features/RushGame/RushGameSections/SelectedCard.tsx";
 import { useBlockNavigation } from "@/hooks/useBlockNavigation.ts";
-import useFetch from "@/hooks/useFetch.ts";
-import useFetchRushBalance from "@/hooks/useFetchRushBalance.ts";
+import { useFetchRushUserParticipationStatus } from "@/hooks/useFetchRushUserParticipationStatus.ts";
 import { useFetchTodayRushEvent } from "@/hooks/useFetchTodayRushEvent.ts";
 import { useRushGameContext } from "@/hooks/useRushGameContext.ts";
 import useToast from "@/hooks/useToast.tsx";
-import { GetRushUserParticipationStatusResponse } from "@/types/rushApi.ts";
 import { writeClipboard } from "@/utils/writeClipboard.ts";
 
 export default function RushGame() {
@@ -25,10 +22,10 @@ export default function RushGame() {
         "이 페이지를 떠나면 모든 변경 사항이 저장되지 않습니다. 페이지를 떠나시겠습니까?"
     );
     const { getTodayRushEvent } = useFetchTodayRushEvent();
-
-    const { gameState, setUserParticipationStatus } = useRushGameContext();
+    const { gameState } = useRushGameContext();
     const { showToast, ToastComponent } = useToast("🔗 링크가 복사되었어요!");
-    const fetchRushBalance = useFetchRushBalance();
+    const { getRushUserParticipationStatus, userParticipatedStatus } =
+        useFetchRushUserParticipationStatus();
 
     const handleClickShareButton = () => {
         writeClipboard(import.meta.env.VITE_RUSH_URL, showToast);
@@ -36,25 +33,8 @@ export default function RushGame() {
 
     useEffect(() => {
         getTodayRushEvent(cookies[COOKIE_KEY.ACCESS_TOKEN]);
-    }, []);
-
-    const { data: userParticipatedStatus, fetchData: getRushUserParticipationStatus } = useFetch<
-        GetRushUserParticipationStatusResponse,
-        string
-    >((token) => RushAPI.getRushUserParticipationStatus(token));
-
-    useEffect(() => {
         getRushUserParticipationStatus(cookies[COOKIE_KEY.ACCESS_TOKEN]);
     }, []);
-
-    useEffect(() => {
-        if (userParticipatedStatus !== null) {
-            setUserParticipationStatus(userParticipatedStatus);
-            if (userParticipatedStatus && CARD_PHASE.IN_PROGRESS) {
-                fetchRushBalance();
-            }
-        }
-    }, [userParticipatedStatus]);
 
     const renderRushGameContent = () => {
         switch (gameState.phase) {
